requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
frozenlist==1.6.0
google-auth==2.39.0
google-genai==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
